improved error checking logic for tileset delete

+++ routes/tilesetRoutes.js
@@ -37,30 +37,25 @@ router.post("/tileset/create", async (req, res) => {
 
 // Delete Tileset
 router.post("/tileset/delete/:id", async (req, res) => {
-  // if (req.session._id == undefined) {
-  //   res.status(400).json({ errorMessage: "Not logged in" });
-  //   return;
-  // }
+  if (req.session._id == undefined) {
+    res.status(400).json({ errorMessage: 'Not logged in' });
+    return;
+  }
   var tileset = await Tileset.findById(req.params.id);
+  if (tileset == null){
+    res.status(400).json({errorMessage: 'Tileset not found'});
+    return;
+  }
   var user_id = tileset.owner;
+  if (user_id != req.session._id){
+    res.status(400).json({errorMessage: 'Not owner of tileset'});
+    return;
+  }
   var user = await User.findById(user_id);
-  // var index = user.tilesets.indexOf(req.params.id);
-  // user.tilesets = user.tilesets.splice(index, 1)
-  // user.save();
   await User.updateOne({ _id: user_id }, { $pullAll: { tilesets: [req.params.id] } });
   await Tileset.findOneAndRemove({ _id: req.params.id })
-    .then(() => res.json({ message: "Tileset deleted" }))
-    .catch((err) => {
-      Tileset.findOne({ _id: req.params.id })
-        .then((tileset) => {
-          if (tileset != null) {
-            res.status(400).json({ errorMessage: "Not owner of tileset" });
-          } else {
-            res.status(400).json({ errorMessage: "Tileset does not exist" });
-          }
-        })
-        .catch((err) => res.status(400).json({ errorMessage: err }));
-    });
+    .then((tileset) => res.json({ message: 'Tileset deleted' }))
+    .catch((err) => { res.status(400).json({ errorMessage: err })});
 });
 
 // Update tileset
